Rework NPC movement to no longer be timer based - Now handled by the 'NPCThinkService' (renamed to 'NpcService'). This prevents insertions and removals of potentially 10 or more 'ArriveAtTargetAction' per second (with a fast 'GAMENPC_FOLLOWCHECK_TIME') per NPC following an object close to them. - Also prevents some waypoint timers from running even on inactive brains (I measured about 15 timers being created and stopped per second by DF mobs, on an empty server. I oddly only noticed it with DF mobs). - While 'FollowTimerCallback' wasn't as bad, stopping and creating a timer just because a NPC aggros something seemed overkill. - Overall reduces the amount of active timers by possibly at least a thousand on a very active server, meaning fewer lock contention in the entity manager and faster iteration in the timer services (at the cost of the NPC service taking a bit longer to check if NPCs have to move). - Most of the movement related code was moved to a new 'MovementComponent', reducing the size of 'GameNPC' a little. - Fixed some odd behavior with pet orders.

+++ GameServer/gameutils/MovementMgr.cs
@@ -16,7 +16,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
  */
-using System.Linq;
+
 using System.Collections.Generic;
 using System.Reflection;
 using DOL.Database;
@@ -158,7 +158,7 @@ namespace DOL.GS.Movement
 	
 				foreach (DBPathPoint pp in pathPoints.Values)
 				{
-					PathPoint p = new PathPoint(pp.X, pp.Y, pp.Z, pp.MaxSpeed, pathType);
+					PathPoint p = new PathPoint(pp.X, pp.Y, pp.Z, (short) pp.MaxSpeed, pathType);
 					p.WaitTime = pp.WaitTime;
 	
 					if (first == null)
